Move classes to individual files: specberus.ts and rule-context.ts
