+++ lib/rules/sotd/pp.ts
@@ -1,7 +1,7 @@
 import type { Cheerio } from 'cheerio';
 import type { Element } from 'domhandler';
 
-import type { RuleContext } from '../../validator.js';
+import type { RuleContext } from '../../rule-context.js';
 import type { RuleCheckFunction, RuleMeta } from '../../types.js';
 
 const self: RuleMeta = {
