+++ lib/rules/sotd/stability.ts
@@ -5,7 +5,7 @@
 import type { Cheerio } from 'cheerio';
 import type { Element } from 'domhandler';
 
-import type { RuleContext } from '../../validator.js';
+import type { RuleContext } from '../../rule-context.js';
 import type { RuleCheckFunction, RuleMeta } from '../../types.js';
 
 async function findSW($candidates: Cheerio<Element>, context: RuleContext) {
