+++ package.json
@@ -3,7 +3,7 @@
   "version": "12.3.1",
   "description": "Specberus is a checker used at W3C to validate the compliance of Technical Reports with publication rules.",
   "license": "MIT",
-  "main": "lib/validator",
+  "main": "lib/specberus",
   "type": "module",
   "repository": {
     "type": "git",
